move typehints to stub
